Extract TypeWalkKeyHelpers.TypeWalkKey; replace 5 identical 3-overload copies (#1065)
Move the shared 3-overload TypeWalkKey (symbol / FQN / FQN+fileLocalDeclaringPath)
into Refactoring.Utilities and retarget ImplementAbstract, ImplementInterface,
GenerateToString, GenerateOverrides, and GenerateEqualsHashCode. Leave
GenerateConstructorOperation's simpler 2-overload alone. Centralize unit tests
in TypeWalkKeyHelpersTests and drop the five duplicate op-level facts.
