+++ tests/RoslynMcp.Core.Tests/Refactoring/Generate/GenerateToStringOperationTests.cs
@@ -3150,37 +3150,6 @@ public class GenerateToStringOperationTests
         Assert.Contains("fields", ex.Message, StringComparison.OrdinalIgnoreCase);
     }
 
-    [Fact]
-    public void TypeWalkKey_IncludesProjectIdentity()
-    {
-        var projectA = ProjectId.CreateNewId();
-        var projectB = ProjectId.CreateNewId();
-        const string fqn = "global::TestApp.Widget";
-
-        var keyA = GenerateToStringOperation.TypeWalkKey(projectA, fqn);
-        var keyB = GenerateToStringOperation.TypeWalkKey(projectB, fqn);
-
-        Assert.NotEqual(keyA, keyB);
-        Assert.Equal(keyA, GenerateToStringOperation.TypeWalkKey(projectA, fqn));
-        Assert.NotEqual(keyA, GenerateToStringOperation.TypeWalkKey(projectA, "global::TestApp.Other"));
-    }
-
-    [Fact]
-    public void TypeWalkKey_FileLocalIdentity_DistinguishesSameFqn()
-    {
-        var project = ProjectId.CreateNewId();
-        const string fqn = "global::TestApp.Worker";
-
-        var ordinary = GenerateToStringOperation.TypeWalkKey(project, fqn);
-        var fileA = GenerateToStringOperation.TypeWalkKey(project, fqn, "/tmp/FileA.cs");
-        var fileB = GenerateToStringOperation.TypeWalkKey(project, fqn, "/tmp/FileB.cs");
-
-        Assert.NotEqual(ordinary, fileA);
-        Assert.NotEqual(ordinary, fileB);
-        Assert.NotEqual(fileA, fileB);
-        Assert.Equal(fileA, GenerateToStringOperation.TypeWalkKey(project, fqn, "/tmp/FileA.cs"));
-        Assert.Equal(ordinary, GenerateToStringOperation.TypeWalkKey(project, fqn));
-    }
 
     [Fact]
     public void Validate_AllFilesTrue_WithLine_Throws()
